# feat: opt for concatenation instead of using fmt.Sprintf