--- a/baggage/baggage_test.go
+++ b/baggage/baggage_test.go
@@ -1029,7 +1029,6 @@ func BenchmarkValueEscape(b *testing.B) {
 }
 
 func BenchmarkMemberString(b *testing.B) {
-	b.ReportAllocs()
 	alphabet := "abcdefghijklmnopqrstuvwxyz"
 	props := make([]Property, len(alphabet))
 	for i, r := range alphabet {
@@ -1037,6 +1036,7 @@ func BenchmarkMemberString(b *testing.B) {
 	}
 	member, err := NewMember(alphabet, alphabet, props...)
 	require.NoError(b, err)
+	b.ReportAllocs()
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_ = member.String()