## Summary
The Member.String() function, a frequently used pathway,
currently relies on `fmt.Sprintf` to merge strings.

However, this method could inadvertently cause additional
memory allocations. By directly concatenating these strings
instead of utilizing fmt and constructing a format template,
we can reduce unnecessary allocations.

This minor modification has the potential to enhance the
efficiency of this library for all users.

## Benchmark strategy
The strategy involves generating a new Member 
with Key, Value and set of properties and call 
`String()`. 

before.txt
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/baggage
BenchmarkMemberString-10         2016388               591.8 ns/op           720 B/op          8 allocs/op
PASS
ok      go.opentelemetry.io/otel/baggage        2.172s
```

after.txt
```
goos: darwin
goarch: arm64
pkg: go.opentelemetry.io/otel/baggage
BenchmarkMemberString-10         2740512               436.4 ns/op           656 B/op          4 allocs/op
PASS
ok      go.opentelemetry.io/otel/baggage        1.989s
```